Tela de pre-cadastro do aluno foi alterada, Tela de pre-cadastro do professor foi feita, Models foram atualizadas

+++ AgendaTCC/database/migrations/2018_04_15_192112_create_tcc_dados_table.php
@@ -16,8 +16,8 @@ class CreateTccDadosTable extends Migration
         Schema::create('tcc_dados', function (Blueprint $table) {
             $table->increments('idDados');
             $table->string('tema', 45);
-            $table->integer('orientador');
-            $table->integer('aluno_matricula');
+            $table->bigInteger('orientador');
+            $table->bigInteger('aluno_matricula');
             $table->string('coorientador', 45);
         });
